w6 final proj update

+++ week6/w6_final_project/src/main.cpp
@@ -45,7 +45,7 @@ int main() {
 			auto predicate = [condition](const Date &date, const string &event) {
 				return condition->Evaluate(date, event);
 			};
-			int count = db.RemoveIf(predicate);
+			size_t count = db.RemoveIf(predicate);
 			cout << "Removed " << count << " entries" << endl;
 		} else if (command == "Find") {
 			auto condition = ParseCondition(is);
